requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2>=2.9.9
djangorestframework-simplejwt>=5.3.1
django-filter>=24.1
django-cors-headers>=4.3.1
# Production requirements
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
python-dotenv>=1.0.1
